+++ tests/telemetry/test_collector.py
@@ -36,6 +36,7 @@ class TestTelemetryCollector:
             "traces": 0,
             "tables": 0,
             "charts": 0,
+            "insights": 0,
             "dashboards": 0,
             "alerts": 0,
             "selectors": 0,
@@ -57,6 +58,7 @@ class TestTelemetryCollector:
         project.alerts = [mock.MagicMock()]
         project.selectors = [mock.MagicMock(), mock.MagicMock()]
         project.destinations = [mock.MagicMock()]
+        project.insights = [mock.MagicMock()]
 
         metrics = collect_project_metrics(project)
 
@@ -70,6 +72,7 @@ class TestTelemetryCollector:
             "alerts": 1,
             "selectors": 2,
             "destinations": 1,
+            "insights": 1,
         }
 
     def test_collect_project_metrics_none_lists(self):
