+++ visivo/models/chart.py
@@ -208,14 +208,7 @@ class Chart(SelectorModel, NamedModel, ParentModel):
 
     def child_items(self):
         """Return child items for DAG construction"""
-        all_items = []
-        if hasattr(self, "traces"):
-            all_items.extend(self.traces)
-        if hasattr(self, "insights"):
-            all_items.extend(self.insights)
-        if hasattr(self, "selector"):
-            all_items.append(self.selector)
-        return all_items
+        return self.traces + self.insights + [self.selector]
 
     traces: List[TraceRef] = Field(
         [],
