+++ visivo/telemetry/collector.py
@@ -24,7 +24,7 @@ def collect_project_metrics(project: Project) -> Dict[str, int]:
         "models": len(project.models) if project.models else 0,
         "traces": len(project.traces) if project.traces else 0,
         "tables": len(project.tables) if project.tables else 0,
-        # "insights": len(project.insights) if project.insights else 0,xs
+        "insights": len(project.insights) if project.insights else 0,
         "charts": len(project.charts) if project.charts else 0,
         "dashboards": len(project.dashboards) if project.dashboards else 0,
         "alerts": len(project.alerts) if project.alerts else 0,
